+++ src/apps/ScopeChangeRequest/Atoms/FormAtomApi/formAtomApi.ts
@@ -57,6 +57,7 @@ export const scopeChangeFormAtomApi = createAtom<ScopeChangeFormModel, FormAtomA
             punchListItemIds: [],
             attachmentsToDuplicate: [],
             revisionAttachments: [],
+            newRevisionOrVoidReason: null,
         }),
 }));